OMAP5: PM: Fix some wrong register offsets
Some register offsets for mainly newly added modules
like gpio7/8, i2c5 etc seems to be generated wrongly.
This might be due to a mismatch in the RTL version
that is used for zebu modeling and the one used
to autogenerate.

While here, also comment out some of the unused
hwmods so they can be uncommented once there is
someone ready to use them.

Signed-off-by: Rajendra Nayak <rnayak@ti.com>
Signed-off-by: sricharan <r.sricharan@ti.com>