OMAP2+: PM: fix watchdog boot when runtime PM is disabled
Fix boot crash in watchdog driver when runtime PM is disabled.

When runtime PM is disabled, devices should be left enabled so that
all device accesses in drivers will succeed even though the runtime PM
get/put calls are noops.

This is already the case for all devices, but the WDT init has its
own hwmod postsetup which was not taking this into account.

This fixes a boot crash where the first device access in the WDT probe
function would crash a booting kernel.

Cc: Paul Walmsley <paul@pwsan.com>
Signed-off-by: Kevin Hilman <khilman@ti.com>