OMAP4: clocks: Add CORE M2 set rate support
This patch adds support for modifying the CORE DPLL M2 value
using the shadow registers. The API programs the CM shadow
registers to update CORE DPLL M2 divider. M2 divider is used to clock
external DDR and its reconfiguration on frequency change is managed
through a hardware sequencer. This is managed by the PRCM with EMIF
using shadow registers.

Currently this is not enabled since it's causing instabilities.

[girishsg@ti.com: Helped in fixing comments]
Signed-off-by: Girish S G <girishsg@ti.com>
Signed-off-by: Rajendra Nayak <rnayak@ti.com>
Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>