usb/compsite: Don't set bcdUSB to 0x0210
I don't know if this is in Windows general or the Virtual Host Controler
that I'm using now but one of those things does not like Device
descriptors with this value.
I keep this in my tree for now and post it to the list in case you might
need it Felipe. The symptoms are: config descriptor fetched, device
descriptor fetched and then silence. Usually Windows would ask for the
string desrcriptor.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>